fix(observability): repair command-center summary/trends/top queries
The Command Center showed no data once telemetry accumulated — the endpoints
errored rather than returning empty:

- summary (500) and top-routes/top-errors: their SUM(value_count) over the
  bigint column returns NUMERIC, which failed to decode into i64. Cast the
  aggregates back to ::bigint. (Invisible on an empty DB: SUM over zero rows
  is NULL, which decodes fine — so it only broke in production with data.)
- trends (400): the endpoint used axum's Query (serde_urlencoded), which
  can't parse repeated `metric=` params into a Vec. Switch to axum_extra's
  Query (serde_html_form); enable the axum-extra `query` feature.
- 7d/30d rollups: the materialized view's SUM columns are also cast to
  ::bigint so the rollup trend query decodes.

Adds an integration regression test that seeds samples and asserts the
summary/top-routes/top-errors/trends queries decode with data present — the
coverage gap that let this ship (previous tests ran against an empty DB).

Verified against production data (types now decode as bigint; migration
populates the matview).

+++ server/src/observability/storage.rs
@@ -411,11 +411,11 @@ pub async fn query_top_routes(
     const BASE: &str = "\
         SELECT \
              labels->>'http.route' AS route, \
-             SUM(value_count) AS request_count, \
+             SUM(value_count)::bigint AS request_count, \
              SUM(CASE \
                  WHEN labels->>'http.response.status_code' ~ '^\\d+$' \
                       AND (labels->>'http.response.status_code')::int >= 500 \
-                 THEN value_count ELSE 0 END) AS error_count, \
+                 THEN value_count ELSE 0 END)::bigint AS error_count, \
              AVG(value_p95) AS avg_p95, \
              MAX(value_p95) AS max_p95 \
          FROM telemetry_metric_samples \
@@ -457,8 +457,8 @@ pub async fn query_top_errors(
     sqlx::query_as::<_, TopRouteEntry>(
         "SELECT \
              labels->>'error.type' AS route, \
-             SUM(value_count) AS request_count, \
-             SUM(value_count) AS error_count, \
+             SUM(value_count)::bigint AS request_count, \
+             SUM(value_count)::bigint AS error_count, \
              AVG(value_p95) AS avg_p95, \
              MAX(value_p95) AS max_p95 \
          FROM telemetry_metric_samples \
